PySpark and Delta Lake
Created a data pipeline that ingests data on Covid19 cases from a Git repository of JHU CSSE and transforms and loads into Databricks DBFS. The idea is to ingest and transform git repo data to a time-series(dates as columns) data such that daily cumulative cases from every country will be stored in confirmed, deaths and recovered case tables and can be used for analysis. The pipeline is adaptive to changing table schema and this is achieved with Delta Lake's Schema Evolution.